+++ driven/storage/adapter.go
@@ -226,8 +226,10 @@ func (sa *Adapter) UpdateOrganization(ID string, name string, requestType string
 	return nil
 }
 
-//GetServiceRegs fetches the requested service registration records
-func (sa *Adapter) GetServiceRegs(serviceIDs []string) ([]authservice.ServiceReg, error) {
+// ============================== ServiceRegs ==============================
+
+//FindServiceRegs fetches the requested service registration records
+func (sa *Adapter) FindServiceRegs(serviceIDs []string) ([]authservice.ServiceReg, error) {
 	var filter bson.M
 	for _, serviceID := range serviceIDs {
 		if serviceID == "all" {
@@ -245,7 +247,44 @@ func (sa *Adapter) GetServiceRegs(serviceIDs []string) ([]authservice.ServiceReg
 		return nil, log.WrapActionError(log.ActionFind, model.TypeServiceReg, &log.FieldArgs{"service_id": serviceIDs}, err)
 	}
 
+	if result == nil {
+		result = []authservice.ServiceReg{}
+	}
+
 	return result, nil
+}
+
+//FindServiceReg finds the service registration in storage
+func (sa *Adapter) FindServiceReg(serviceID string) (*authservice.ServiceReg, error) {
+	filter := bson.M{"service_id": serviceID}
+	var reg *authservice.ServiceReg
+	err := sa.db.serviceRegs.FindOne(filter, &reg, nil)
+	if err != nil {
+		return nil, log.WrapActionError(log.ActionFind, model.TypeServiceReg, &log.FieldArgs{"service_id": serviceID}, err)
+	}
+
+	return reg, nil
+}
+
+//InsertServiceReg inserts the service registration to storage
+func (sa *Adapter) InsertServiceReg(reg *authservice.ServiceReg) error {
+	_, err := sa.db.serviceRegs.InsertOne(reg)
+	if err != nil {
+		return log.WrapActionError(log.ActionInsert, model.TypeServiceReg, &log.FieldArgs{"service_id": reg.ServiceID}, err)
+	}
+
+	return nil
+}
+
+//UpdateServiceReg updates the service registration in storage
+func (sa *Adapter) UpdateServiceReg(reg *authservice.ServiceReg) error {
+	filter := bson.M{"service_id": reg.ServiceID}
+	err := sa.db.serviceRegs.ReplaceOne(filter, reg, nil)
+	if err != nil {
+		return log.WrapActionError(log.ActionInsert, model.TypeServiceReg, &log.FieldArgs{"service_id": reg.ServiceID}, err)
+	}
+
+	return nil
 }
 
 //SaveServiceReg saves the service registration to the storage
@@ -260,11 +299,29 @@ func (sa *Adapter) SaveServiceReg(reg *authservice.ServiceReg) error {
 	return nil
 }
 
+//SaveServiceReg deletes the service registration from storage
+func (sa *Adapter) DeleteServiceReg(serviceID string) error {
+	filter := bson.M{"service_id": serviceID}
+	result, err := sa.db.serviceRegs.DeleteOne(filter, nil)
+	if err != nil {
+		return log.WrapActionError(log.ActionDelete, model.TypeServiceReg, &log.FieldArgs{"service_id": serviceID}, err)
+	}
+	if result == nil {
+		return log.WrapDataError(log.StatusInvalid, "result", &log.FieldArgs{"service_id": serviceID}, err)
+	}
+	deletedCount := result.DeletedCount
+	if deletedCount == 0 {
+		return log.WrapDataError(log.StatusMissing, model.TypeServiceReg, &log.FieldArgs{"service_id": serviceID}, err)
+	}
+
+	return nil
+}
+
 //NewStorageAdapter creates a new storage adapter instance
 func NewStorageAdapter(mongoDBAuth string, mongoDBName string, mongoTimeout string, logger *log.Logger) *Adapter {
 	timeoutInt, err := strconv.Atoi(mongoTimeout)
 	if err != nil {
-		logger.Error("Set default timeout - 500")
+		logger.Warn("Setting default Mongo timeout - 500")
 		timeoutInt = 500
 	}
 	timeout := time.Millisecond * time.Duration(timeoutInt)
@@ -278,18 +335,15 @@ func abortTransaction(sessionContext mongo.SessionContext) {
 	if err != nil {
 		//TODO - log
 	}
-
 }
 
 //Listener represents storage listener
 type Listener interface {
 	OnAuthConfigUpdated()
+	OnServiceRegsUpdated()
 }
 
-//DefaultListenerImpl represents default storage listener implementation
-type DefaultListenerImpl struct {
-}
+type DefaultListenerImpl struct{}
 
-//OnAuthConfigUpdated notifies that the auth config has been updated
-func (d *DefaultListenerImpl) OnAuthConfigUpdated() {
-}
+func (d *DefaultListenerImpl) OnAuthConfigUpdated()  {}
+func (d *DefaultListenerImpl) OnServiceRegsUpdated() {}